Fix and simplify handling of `:` method calls
Stop storing them as a special `@methcall` node type, and instead store a `:`
binary node. Given expression `a:b`, the `:` node is typed with the type of
`a.b`. In the handling for `@funcall`, we detect a `:` child in `e1` and then
insert the implicit "self" first argument.

tl tl.tl: 2 errors