Introduce cast operator `as`
Minimal implementation of a cast operator (`EXP as TYPE`).

With this, we are able to annotate the dynamic trickery done when
copying the type table, and we reach the mark of 0 errors as the
type checker typechecks itself!

Due to my laziness in adjusting the parser, the operator currently
applies only to top-level expressions. This needs to be eventually
fixed.

The operator could probably be made somewhat safer by only allowing casts
between table types, but a general unsafe operator may come handy when
metatables introduce funny overloaded behavior.

tl tl.tl: 0 errors