enums
Add support for `enum` type declarations.

An `enum` is a string that can only be assigned to a literal
from a specific set.

An `enum` can be assigned to a variable of type `string`.
A `string` can only be assigned to a variable of an enum type
if it is a literal declared in that an enum.

Enums with the same contents but different names are considered to be
different (i.e. they are checked nominally, not structurally).