Remove dependency on lua_no_tailcalls
`lua_no_tailcalls` is my custom build of the Lua interpreter which I use for
development. The only difference is that the interpreter never emits the
TAILCALL bytecode, emitting always plain CALL.

This of course removes tail call support from Lua (which is not an
optimization, but rather a guarantee of feasibility for recursive algorithms
that use it), but on the other hand produces full stack traces useful for
debugging.

Since this is not available on everyone's machines, I'm removing all references
to it from this repo.