Add support for loading .d.tl files as definition files
When type checking, `require("foo")` will search for:

* `foo.tl` first
* then `foo.d.tl`
* and finally `foo.lua`

This means that if a `.d.tl` definitions file exists,
it will be used as a stand-in for a Lua module, and the
`.lua` file contents will be ignored when typechecking.

When running code, we do not read `.d.tl` files, and just check

* `foo.tl` first
* and finally `foo.lua`

This means that if a Lua module has a pair of files `foo.d.tl`
and `foo.lua`, the `.d.tl` file will be used for typechecking
and the `.lua` file will be used for running. It is the user's
responsibility to ensure that these two files match -- for now
we perform no cross-checking between them.