Fix propagation of unknown values and function returns
When Lua functions are not annotated, tl doesn't know much about them: it
doesn't know the function argument types, and it doesn't know the number of
returns. We don't do flow analysis (yet?) to try to discover that based on the
`return` statements -- that can get pretty complicated and still unable to get
firm results because `return f()` where `f` is a function coming from a
different module can return any number of values... so we better not put
too much effort on this kind of guesswork and instead encourage programmers
to annotate their return types.

Considering the test case reported by @mwild1:

```
local data = f1().data
one, two, three = data()
return one, two, three
```

What's happening here is that tl doesn't really know the return type of `f1`.
For that reason, it also doesn't know the type of `data`. In that situation,
what we need to do is to propagate the unknown state: when `data` is used as a
function, we assume that it is a function that is vararg-unknown in,
vararg-unknown out. Therefore we shouldn't produce _errors_ to the assignment
of its return values, but instead report them as unknown. When the user
finally annotates enough variables so that a known state propagates to those
variables, tl will be able to tell if their usage is correct or not.

Fixes #55.