Fix resolution of generic types by delaying resolution of typeargs
Adds a distinction between types being "found" and "resolved".

A type `T<A>` is "found" when its base type `T` is found in scope.

A concrete type `T<x>` which is an instance of `T<A>` is "resolved"
when both its base type `T` is found in scope and the type argument
`x` is also found in scope. Resolution produces a new type object
(stored in the `.resolved` field of the nominal object) which is
a copy of `T` where all instances of type variable `A` are
replaced with `x`.

For non-generic types, a type being "found" and "resolved" means
the same.

Fixes #167.