tl: add build options and tl build command
* Adds luafilesystem dependency and the following config options:

- build_dir
- source_dir
- files
- include
- exclude
- Original include -> include_dir

* tl build: new command

* tl: completely ignore .d.tl files when traversing directory.

* tl run: exit when load/loadstring reports error

* tl gen will now always put the generated file in the current dir,
unless specified with -o (not implemented yet)

Tests:

* Add initial tests for globbing

In particular adds the following util function: util.write_tmp_dir,
which lets us build up a directory structure in /tmp to test out tl gen.

The run_mock_project will probably move to util as more tests get
written, but in short: it uses util.write_tmp_dir then makes a link to
tl and tl.lua in the dir, does a lfs.chdir, and runs ./tl gen. Then it
asserts that the generates files are as expected, and cleans up after
itself by removing the dir and does a lfs.chdir back to the correct dir

* Add initial build dir tests

The run_mock_project was moved into utils to help with this and given
some new functionality. It can now specify the command to run and the
arguments to run it with. As well as some basic assertions about the
types of its arguments.

Internal changes:

* I've attempted to separate each command into its own section of the
  file where it will do its thing and then exit. This favors build to have
  the last chunk of the file all to itself to build up a source map and
  all sorts of data structures

* build is now a lot more organized in how it handles
  the directory. In particular paths are now all relative and there is no
  more messy juggling of absolute vs relative paths. In addition, the
  'project' table/object is a good candidate for an api as it provides
  a neat representation of the project directory as well as some methods
  to navigate it.

Thanks to @pdesaulniers for the help when applying review comments.