spec: Add initial tests for globbing
In particular adds the following util function: util.write_tmp_dir,
which lets us build up a directory structure in /tmp to test out tl gen.

The run_mock_project will probably move to util as more tests get
written, but in short: it uses util.write_tmp_dir then makes a link to
tl and tl.lua in the dir, does a lfs.chdir, and runs ./tl gen. Then it
asserts that the generates files are as expected, and cleans up after
itself by removing the dir and does a lfs.chdir back to the correct dir