tl: Huge refactors + add build command
With this commit a lot of tests had to be put as pending to be
managable.

External changes:
- tl gen was reverted to its old behavior in favor of a new tl build command.
- tl gen will now always put the generated file in the current dir,
unless specified with -o (not implemented yet)
- tl build now mimics the behavior of the previous tl gen

Internal changes:
- I've attempted to separate each command into its own section of the
file where it will do its thing and then exit. This favors build to have
the last chunk of the file all to itself to build up a source map and
all sorts of data structures
- build (previously gen) is now a lot more organized in how it handles
the directory. In particular paths are now all relative and there is no
more messy juggling of absolute vs relative paths. In addition, the
'project' table/object is a good candidate for an api as it provides
a neat representation of the project directory as well as some methods
to navigate it.