fix: do not shadow argument types with function arguments
This includes a big refactor, because begin_function_scope used to traverse
the sub-nodes of the function declaration on its own, instead of letting the
visitor pattern from recurse_node and recurse_type do it. Now, the traversal
logic for function declarations is controlled by the visitor, and as a result
some annotations in the AST were fixed to better distinguish between
variables, arguments and identifiers, so the visitor can do its job.