improve "X is not a X" messages
Sometimes different types will have the same name, and we don't
have a canonical full-path name, because modules can be required
under different names, so `modulename.Type` is not always a valid
representation of type `Type` declared in module `modulename`.

This will show the filename and line number of the declared types
when two mismatched types have identical names.