fix: make tl.process always return a new Result
This makes tl.process_string and tl.process always return a new Result object.
If an existing Result is passed as input, the new Result shares the lists of
errors with the old one (so that errors accumulate, as expected by the main tl
driver and the tests), but the `type` and `ast` fields become unique per
processed file.

Fixes #265.