Improve inference for tuples
To convert a tuple to an array, first try a basic union of all the types
in the tuple. Failing that, try to expand the types.
This fixes the issues with things like table.unpack as well as ipairs
since tuples are now much easier to be converted to arrays