Change arrays/maps are inferred when literal number indexes are used
The main change with this is that if a table literal has a `[number] =
value` field, unless the number is a positive integer literal, the table
now gets inferred as a map with number keys. This conflicts with how it
works previously and with some of the tests checking for array-record
inference.

Additionally, this fixes an uncaught bug in both arrays and tupletables
where explicit number indexes would be counted as implicit array
indexes. So a table like `{ [-1] = "str" , [-2] = 4 }` was allowed to be
assigned to both an array `{string | number}` and tuple `{string, number}`
since they were counted as the implicit array entries.

(also, maybe type tuples internally should be called something else?
tupletables is a little silly, but what is used internally to
distinguish them. Maybe something like 'multi'?)