Allow some array literals in assignment to heterogenous tuples
Previously arrays would outright be incompatible due to the entire tuple
being compared, but since the type of array literals now carry their
length, we now only check up to that length for compatability (provided
that it's positive*)

e.g. Assignments like `local t: {number, string}
= { 1 }` are now valid, where they previously weren't

More formally: One can imagine an array as an infinite tuple of a single
type. To compare an array to a tuple is to truncate the array to a tuple
of matching size, then compare component-wise. If that array has a
known/inferred length, then both tuples can be truncated to that length.

* Arguably "0 length" arrays could be treated as length 1 as the sort of
  'minimum' requirement for an array to match a tuple is at least the
  first element must match

Co-authored-by: Hisham Muhammad <hisham@gobolinux.org>