ensure that prefixexp constraints are honored by the parser
In function calls, there are Lua restrictions on what can
go in the lhs. We were not checking this before and we we were
accepting arbitrary expressions at the parser level, and then
catching invalid lhs expressions at the type checking stage.

However, this lets some invalid Lua through for non-parenthesized exps.
So let's catch it at parse time.