drop support for record and enum definitions inside tables
This is a remnant from when variable and type definitions were mixed. Let's
switch to keeping types (with `local type T = D`, etc.) and variables (with
`local V: T = E`) separate.

A run of this branch through `teal-types` showed that no definitions relied
on that syntax, and the diff in `require` shows that it's easy to change it.