fix type inference for variadic return in for loop
Fixes #237.

The root of the problem in the test case from #237 (added here as
a regression test) was that when the type variable `T` in
the return type `T...` was being resolved, the `is_va` attribute of
the type vas being dropped, and would turn into a non-variadic
concrete type `string`.

My fix for this was to rework the internals to that the `is_va` attribute
became an attribute of the _tuple_ instead of the type variable. This
way, the return `(T...)` cleanly converts into `(string...)`.

The PR looks pretty clean, but this was pretty tricky to refactor, because of
the conflation between `{Type}` and `Type` (due to `Type` being an
array-record), so tracking down uses of tuples and plain `{Type}` arrays
(which are sometimes converted into each other) is sometimes confusing in the
codebase.

(I have considered making `Type` no longer be an array-record, but that might
require a more drastic rework of the tree-walking code for type-checking and
code-gen, both of which assume that `Node` and `Type` are tree-like
structures. Also, Teal we had tagged records and all tuples had been properly
identified as their own type throughout the code, that would have been helpful
too!)