do not emit a warning when localizing a global
Localizing globals is a common Lua pattern (for better or worse...)
so let's not emit a warning when we detect that it is happening.

For simplicity, I don't check that the variable being aliased with
the same name is really a global, but that shouldn't matter.

This does not check that the dynamic _value_ is the same, only that the static
_name_ is the same. Therefore, people shouldn't do this to localize the
standard library: `local os = require("os")` but instead do this `local os =
os` (though it's often not needed because the Teal compat53 code generator
already does that for the standard library modules in use).

See #285.