Add --gen-target flag, instead of relying on the Lua version running the compiler
Avoid relying on auto-detection based on the Lua version running the compiler.
This could cause problems, for example, if we ship a binary bundle built with
Lua 5.4 and someone wants to build code to run with LuaJIT.

We still use version auto-detection as a fallback, which is useful for the case
when the Teal compiler is used as a package loder (and therefore runs on the
same VM where the Teal module being compiled will run).

Also improve docs for --gen-compat, and deprecate --skip-compat53 (which is
no longer advertised, but is still supported).