Fix casting empty expressions + explicit failure messages
Previously expressions like `() as string` and `1 as` would crash the
compiler, these now produce syntax errors instead.

We only produce a bare "syntax errors" in truly ambiguous situations,
such as when there are many possibilities and an invalid token was
seen. Note that this means "syntax error" in functions such as
parse_call_or_assigment and parse_literal because those happen
to be the last option tried in a statement or expression context,
so any invalid token ends up falling through to these functions
(i.e. it runs in situations where not specifically a literal, call
or assigment is expected — in this situations it's best just to
point to the error location and let the user figure out the typo).

Additionally, any call to `fail` now must have an associated error
message.

This commit also adds more specific error messages for malformed strings
and numbers.

Co-Authored-By: Hisham Muhammad <hisham@gobolinux.org>