tl.tl: Include a tk: string field in tl.Error + expose Token
Most of the token information was already present in tl.Error, now just
include the string as well for tooling like cyan, so they don't have to
search for the token to get the length/content of it.

Also expose Token and TokenKind as tl.Token and tl.TokenKind since they
are relatively simple types.