tl: script_path code moved around, move back to the front
This was originally added in 2019 to be at the beginning of the
file, for the purpose explained in the comment below. Eventually
it ended up moving around in the file, which made it ineffective.
We've had reports of users confused that a git clone of tl was picking
up the wrong tl.lua, so this little trick seems to have value --
let's bring it back.