tests: do not modify Type tables
Modifying Type tables returned as parts of the AST can break
the compiler (for example, the invariant here that every Type
has a typeid is broken for the following tests).

For a public AST API, we'll need to ensure we return copies
of the produced types, and not the actual Type records used
by the compiler as part of the standard library definition
or in definitions such as INVALID.

For now, we'll just be careful in the tests.