canonicalize pathnames to improve behavior of env.loaded
This does not go as far as converting pathnames to absolute
as that would require going beyond pure-Lua, but it does
improve the situation. In particular, it should allow
type-checking the global_env_def file from the current folder
without a confusion between "file.d.tl" and "./file.d.tl".

Fixes #518.