new attribute: <total>
The `<total>` attribute declares a const variable with a table
where the domain of its keys are totally declared.

This check can only be applied when a literal table is given
at the time of variable initialization, and only for table types
with well-known finite domains: maps with enum keys, maps with
boolean keys, records.

Note that the requirement is that keys are declared: they may
still be explicitly declared to be nil.