interfaces/records: move arrayrecord to `is {...}`
This avoids a grammar ambiguity in the `where` clause, where a following `{T}`
would be mistaken as being part of the `where` expression.

We still support a `{T}` at the top of the record declaration if there is no
`is` clause, but not in the middle of the field list anymore.