standard library: support "n"/"*n" in file reader functions
A pragmatic implementation which describes it as a polymorphic function
which covers the most common cases first, avoiding the need for casts.

Yes, this is a form of poor-man's dependent typing and abusing enums
as poor-man's literal types. No, I don't want to add fixed-arity cases
for all combinations of 2, 3, 4... arguments -- 1-arity and the variadic
fallback should be enough.

Closes #718.