avoid setting (or updating!) typename directly
This is to ensure that there is always a consistent correspondence between a
typeid and its typename (that is, if the typename changes, we update the
typeid as well).

With consistency between typeids and typenames, we would _almost_ be able to
cache type comparison judgements based on typeids alone. We're not able to do
it just because type comparisons such as is_a and same_type have one possible
side-effect: inferring typevars.

From my experiments, in the current state of the code there's no real
performance gain by caching type comparisons, but it would be nice to be able
to do it for better understandability -- intuitively, type comparisons should
be a pure operation.

In any case, the changes in this commit should make it easier to reason about
and debug the code in the long term.