stricter checks for shadowing of type arguments
Produces an error when using a base type as a type argument,
and a warning when using another variable name. We can be
stricter with type arguments because those exist only in
Teal's type world. We can't be as strict with redeclarations
in general because record types also exist as concrete Lua
objects (example: the "string" table in the standard library).

Fixes #764.