metatables: check metamethod types in metatable definition
Add special-case behavior to specialize a type `metatable<R>` using the
definition of `metamethod` entries from `R` (and not just a type-variable
application of `R` into the definition of `global record metatable<T>` from
the standard library definition.

See tests in spec/declaration/metatable_spec.lua for examples of the added
checks.

Fixes #633.

(At least the extent of it that can be resolved at this time,
without explicit `nil` support -- a good explanation as to why
the second case isn't resolved is given by @bjornbm in
https://github.com/teal-language/tl/issues/633#issuecomment-1450731594 :

"the record definition defines what keys/values are valid, but not that they
are defined (or more generally perhaps the values may be nil, since nil is a
valid value of every type). What is checked is that values for the defined
keys have the right type, and that no other keys are added to the record.