string.gsub: accept functions with no returns
Note that we're doing a little hack here, but notating the
return type as vararg, as a stand-in for something like a `?`
in the return type. I guess we need `min_arity` calculations
in the return types as well.

Also fix the declaration of `gsub` because Lua does not accept
functions that return booleans, but it does accept numbers.