avoid "got foo, expected foo" errors
Do not produce a confusing message when reporting on incompatible types with
the same local names. Before this commit, the included test case would produce
an error message saying "got X | integer, expected X | integer"; which was
technically correct because it referred to different nested records called X,
but it was confusing. Since we don't "resolve path names" based on the
position of the error, bailing out with a generic error message is better than
producing a confusing one.