refactor: restrict special_function_handler to FunctionType
One of the main sources of maintenance headaches for Teal
in the past was being able to track the various attributes
of types, that is, determining under which circumstances
each field could be set and under which ones they couldn't.

A big project in 2024 was to split the big Type record type
into several types and interfaces, one for each type. We now
avoid having type-specific behaviors defined in the more
general types.

This doesn't look super pretty either (especially with storing
the handler in the first function of the poly type), but it
keeps the type hierarchy cleaner at the expense of a bit more
type-check-time logic (`resolve_for_special_function`).