simplify prelude and stdlib generation
They are now plain .d.tl files, and in `new_env` (which is
currently the "manual" environment builder), they are loaded using the
standard require_module mechanism, same as the other predefined_modules
(or any other uses of `require` by Teal code, really!)

Of course, `environment.default` now uses the precompiled one.

And yes, this is an experiment in the direction of perhaps someday
providing a precompiled type definition facility for all modules,
aiming to speed up compilations. This has the difficulty that the
Teal environment currently relies on globally unique numeric typeids.
We'd have to switch to a name-based system with a linking step.
(The standard library can get away with this because it is always
the first module on an empty environment, so I can reliably start
my typeid count from 1.)