refactor: simplify handling of gen_target
We rollback to checking the '<close>' attribute during check
(currently the only gen_target-related check we make).

We only *modify* the AST for compat changes in `teal.gen.lua_compat`.

This allows us to assume in `teal.gen.*` that the AST is correct
and no further errors need to be produced.