refactor: simplify handling of environment options
We no longer support overriding the environment options
in our check() API. This simplifies the propagation of
options a lot.

In practice, I don't think this feature was ever used:
both `tl` and `cyan` send the same values to the environment
and to `check()`. In the `v2` API handler, we now just
verify that these values are consistent, since the
internals no longer handle inconsistencies between them.