make feat_lax always auto-detected
Make it always decide it at parse time, based on filename extension,
or, failing that, the hashbang heuristic.

The `tl` CLI did not expose setting the lax mode explicitly anyway, and Cyan
always sets its consistently with the environment. So, we simplify all APIs
since `.lua` files without lax mode are not meant to be a supported
configuration: when not in lax mode, those files should be `.tl` files anyway:
we don't want to further complicate the number of dialects a `.lua` file can
have (and due to our efforts with compat handling, we want `.tl` files to be
as consistent as possible no matter which Lua VM is used underneath).