Implement Lua 5.5 style named variadic arguments
Lua 5.5 allows an identifier after a `...` to collect variadic arguments
into a table, this patch implements that for Teal by generating that
collection manually. Like so:

```teal
local function foo(...ints: integer)
	local x: {integer} = ints
	local y = select(2, ...) -- Note: ... is still usable
end
```

```lua
local function foo(...)local ints={...};
	local x = ints
	local y = select(2, ...)
end
```

The main part of this patch is unifying function generation between
anonymous, record, local, and global functions to handle this in a
central location.

The secondary part in `visitors.tl` ensures to callers of the function
that it is variadic, but in the body, the bound name is an array.

To achieve this, when parsing a "..." in an argument list followed by an
identifier, that identifier node is stored in the `name` field of the
"...".